requirements.txt
fastapi==0.115.5
uvicorn==0.30.6
SQLAlchemy==2.0.36
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.9.2
python-dotenv==1.0.1
email-validator==2.2.0